move rest of legacy events, fragment ext, fix imports
