+++ android/src/main/java/com/swmansion/rnscreens/legacy/Screen.kt
@@ -26,8 +26,6 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.shape.CornerFamily
 import com.google.android.material.shape.MaterialShapeDrawable
 import com.google.android.material.shape.ShapeAppearanceModel
-import com.swmansion.rnscreens.events.SheetDetentChangedEvent
-import com.swmansion.rnscreens.ext.asScreenStackFragment
 import com.swmansion.rnscreens.gamma.common.FragmentProviding
 import com.swmansion.rnscreens.legacy.bottomsheet.SheetAnimationCoordinator
 import com.swmansion.rnscreens.legacy.bottomsheet.SheetDetents
@@ -38,6 +36,8 @@ import com.swmansion.rnscreens.legacy.bottomsheet.updateMetrics
 import com.swmansion.rnscreens.legacy.bottomsheet.useSingleDetent
 import com.swmansion.rnscreens.legacy.bottomsheet.usesFormSheetPresentation
 import com.swmansion.rnscreens.legacy.events.HeaderHeightChangeEvent
+import com.swmansion.rnscreens.legacy.events.SheetDetentChangedEvent
+import com.swmansion.rnscreens.legacy.ext.asScreenStackFragment
 import com.swmansion.rnscreens.utils.pxToDp
 import kotlin.math.max
