+++ android/src/main/java/com/swmansion/rnscreens/legacy/ScreenStackHeaderConfig.kt
@@ -502,7 +502,7 @@ class ScreenStackHeaderConfig(
 
         // set primary color as background by default
         val tv = TypedValue()
-        if (context.theme.resolveAttribute(R.attr.colorPrimary, tv, true)) {
+        if (context.theme.resolveAttribute(android.R.attr.colorPrimary, tv, true)) {
             toolbar.setBackgroundColor(tv.data)
         }
         toolbar.clipChildren = false
