+++ android/src/main/java/com/swmansion/rnscreens/legacy/ScreenViewManager.kt
@@ -12,9 +12,6 @@ import com.facebook.react.uimanager.ViewGroupManager
 import com.facebook.react.uimanager.ViewManagerDelegate
 import com.facebook.react.viewmanagers.RNSScreenManagerDelegate
 import com.facebook.react.viewmanagers.RNSScreenManagerInterface
-import com.swmansion.rnscreens.events.ScreenWillAppearEvent
-import com.swmansion.rnscreens.events.ScreenWillDisappearEvent
-import com.swmansion.rnscreens.events.SheetDetentChangedEvent
 import com.swmansion.rnscreens.legacy.bottomsheet.SheetDetents
 import com.swmansion.rnscreens.legacy.events.HeaderBackButtonClickedEvent
 import com.swmansion.rnscreens.legacy.events.HeaderHeightChangeEvent
@@ -22,6 +19,9 @@ import com.swmansion.rnscreens.legacy.events.ScreenAppearEvent
 import com.swmansion.rnscreens.legacy.events.ScreenDisappearEvent
 import com.swmansion.rnscreens.legacy.events.ScreenDismissedEvent
 import com.swmansion.rnscreens.legacy.events.ScreenTransitionProgressEvent
+import com.swmansion.rnscreens.legacy.events.ScreenWillAppearEvent
+import com.swmansion.rnscreens.legacy.events.ScreenWillDisappearEvent
+import com.swmansion.rnscreens.legacy.events.SheetDetentChangedEvent
 
 @ReactModule(name = ScreenViewManager.REACT_CLASS)
 open class ScreenViewManager :
